Hikari PolyBot/requirements.txt
hikari
hikari-lightbulb
lavasnek_rs

python-dotenv
dnspython
aiohttp
cairosvg
chess
matplotlib
motor
pynacl
